Ft: update and delete role logic

+++ app/Http/Controllers/Back/Authorization/RoleController.php
@@ -9,27 +9,6 @@ use Illuminate\Http\Request;
 
 class RoleController extends Controller
 {
-    /**
-     * Check if authenticated admin has the proper permissions. If not
-     * a redirect route is given back to direct the user to
-     *
-     * @param $permissions
-     * @return bool|\Illuminate\Http\RedirectResponse
-     */
-    protected function redirectIfCant($permissions)
-    {
-        if(admin()->cant($permissions)) {
-            return redirect()->route('back.dashboard')->with('messages.error', 'Oeps. Het lijkt erop dat je geen toegang hebt tot dit deel van chief. Vraag even de beheerder voor meer info.');
-        }
-
-        return false;
-    }
-
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function index()
     {
         $roles = Role::all();
@@ -38,7 +17,7 @@ class RoleController extends Controller
 
     public function create()
     {
-        if($redirect = $this->redirectIfCant('create-role')) return $redirect;
+        $this->authorize('create-role');
 
         $permissions = Permission::all();
         return view('back.authorization.roles.create', ['permissions'=>$permissions]);
@@ -46,7 +25,7 @@ class RoleController extends Controller
 
     public function store(Request $request)
     {
-        if($redirect = $this->redirectIfCant('create-role')) return $redirect;
+        $this->authorize('create-role');
 
         $this->validate($request, [
             'name'               => 'required|unique:roles',
@@ -59,69 +38,43 @@ class RoleController extends Controller
         return redirect()->route('back.roles.index')
                          ->with('messages.success', 'Rol '. $role->name.' is toegevoegd.');
     }
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        return redirect('roles');
-    }
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function edit($id)
     {
+        $this->authorize('update-role');
+
         $role = Role::findOrFail($id);
         $permissions = Permission::all();
-        return view('back.roles.edit', compact('role', 'permissions'));
+
+        return view('back.authorization.roles.edit', compact('role', 'permissions'));
     }
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function update(Request $request, $id)
     {
-        $role = Role::findOrFail($id);
+        $this->authorize('update-role');
+
         $this->validate($request, [
-            'name'=>'required|max:10|unique:roles,name,'.$id,
-            'permissions' =>'required',
+            'name' => 'required|unique:roles,name,'.$id,
+            'permission_names' =>'required|array',
         ]);
-        $input = $request->except(['permissions']);
-        $permissions = $request['permissions'];
-        $role->fill($input)->save();
-        $p_all = Permission::all();
-        foreach ($p_all as $p) {
-            $role->revokePermissionTo($p);
-        }
-        foreach ($permissions as $permission) {
-            $p = Permission::where('id', '=', $permission)->firstOrFail(); //Get corresponding form permission in db
-            $role->givePermissionTo($p);
-        }
+
+        $role = Role::findOrFail($id);
+        $role->name = $request->name;
+        $role->save();
+        $role->syncPermissions($request->permission_names);
+
         return redirect()->route('back.roles.index')
-            ->with('flash_message',
-                'Role'. $role->name.' updated!');
+            ->with('messages.success', 'Rol '. $role->name.' is aangepast.');
     }
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function destroy($id)
     {
+        $this->authorize('delete-role');
+
         $role = Role::findOrFail($id);
         $role->delete();
+
         return redirect()->route('back.roles.index')
-            ->with('flash_message',
-                'Role deleted!');
+            ->with('messages.success', 'Rol '. $role->name.' is verwijderd.');
     }
 }
